Remove BUILD_TEST dependance on BUILD_IMPORT
