+++ cMake/FreeCAD_Helpers/CheckInterModuleDependencies.cmake
@@ -42,5 +42,4 @@ macro(CheckInterModuleDependencies)
     REQUIRES_MODS(BUILD_SPREADSHEET        BUILD_DRAFT)
     REQUIRES_MODS(BUILD_SURFACE            BUILD_PART)
     REQUIRES_MODS(BUILD_TECHDRAW           BUILD_PART BUILD_PART_DESIGN BUILD_SPREADSHEET BUILD_MEASURE BUILD_IMPORT)
-    REQUIRES_MODS(BUILD_TEST               BUILD_IMPORT)
 endmacro(CheckInterModuleDependencies)
